move model::property module into road_network
