+++ rust/routee-compass-core/src/algorithm/search/util/edge_cut_frontier_model.rs
@@ -1,6 +1,8 @@
+use crate::model::{
+    frontier::frontier_model::FrontierModel,
+    road_network::{Edge, EdgeId},
+};
 use std::{collections::HashSet, sync::Arc};
-
-use crate::model::{frontier::frontier_model::FrontierModel, road_network::edge_id::EdgeId};
 
 /// A wrapper of the user-generated FrontierModel which prohibits traversals
 /// on selected edges. algorithms can create this wrapper with a set of "cut edges"
@@ -25,9 +27,9 @@ impl EdgeCutFrontierModel {
 impl FrontierModel for EdgeCutFrontierModel {
     fn valid_frontier(
         &self,
-        edge: &crate::model::property::edge::Edge,
+        edge: &Edge,
         state: &[crate::model::traversal::state::state_variable::StateVar],
-        previous_edge: Option<&crate::model::property::edge::Edge>,
+        previous_edge: Option<&Edge>,
         state_model: &crate::model::state::state_model::StateModel,
     ) -> Result<bool, crate::model::frontier::frontier_model_error::FrontierModelError> {
         if self.cut_edges.contains(&edge.edge_id) {
