+++ rust/routee-compass-core/src/model/access/default/turn_delays/turn_delay_access_model.rs
@@ -1,7 +1,7 @@
 use super::turn_delay_access_model_engine::TurnDelayAccessModelEngine;
 use crate::model::{
     access::{access_model::AccessModel, access_model_error::AccessModelError},
-    property::{edge::Edge, vertex::Vertex},
+    road_network::{Edge, Vertex},
     state::{state_feature::StateFeature, state_model::StateModel},
     traversal::state::state_variable::StateVar,
 };
