+++ rust/routee-compass-core/src/model/frontier/frontier_model.rs
@@ -1,7 +1,6 @@
 use super::frontier_model_error::FrontierModelError;
 use crate::model::{
-    property::edge::Edge, state::state_model::StateModel,
-    traversal::state::state_variable::StateVar,
+    road_network::Edge, state::state_model::StateModel, traversal::state::state_variable::StateVar,
 };
 
 /// Validates edge and traversal states. Provides an API for removing edges from
