+++ rust/routee-compass-core/src/model/mod.rs
@@ -1,7 +1,6 @@
 pub mod access;
 pub mod cost;
 pub mod frontier;
-pub mod property;
 pub mod road_network;
 pub mod state;
 pub mod termination;
