--- rust/routee-compass-core/src/model/property/mod.rs
@@ -1,2 +1,0 @@
-pub mod edge;
-pub mod vertex;
