+++ rust/routee-compass-core/src/model/road_network/mod.rs
@@ -1,7 +1,16 @@
+pub mod edge;
 pub mod edge_id;
 pub mod edge_loader;
 pub mod graph;
 pub mod graph_error;
 pub mod graph_loader;
+pub mod vertex;
 pub mod vertex_id;
 pub mod vertex_loader;
+
+pub use edge::Edge;
+pub use edge_id::EdgeId;
+pub use graph::Graph;
+pub use graph_error::GraphError;
+pub use vertex::Vertex;
+pub use vertex_id::VertexId;
