+++ rust/routee-compass-core/src/model/road_network/test/vertices.csv
