+++ rust/routee-compass-powertrain/src/routee/energy_traversal_model.rs
@@ -1,8 +1,7 @@
 use super::energy_model_ops::get_grade;
 use super::energy_model_service::EnergyModelService;
 use super::vehicle::vehicle_type::VehicleType;
-use routee_compass_core::model::property::edge::Edge;
-use routee_compass_core::model::property::vertex::Vertex;
+use routee_compass_core::model::road_network::{Edge, Vertex};
 use routee_compass_core::model::state::state_feature::StateFeature;
 use routee_compass_core::model::state::state_model::StateModel;
 use routee_compass_core::model::traversal::state::state_variable::StateVar;
