+++ rust/routee-compass/src/app/compass/config/frontier_model/combined/combined_model.rs
@@ -1,6 +1,6 @@
 use routee_compass_core::model::{
     frontier::{frontier_model::FrontierModel, frontier_model_error::FrontierModelError},
-    property::edge::Edge,
+    road_network::Edge,
     state::state_model::StateModel,
     traversal::state::state_variable::StateVar,
 };
