+++ rust/routee-compass/src/app/compass/config/frontier_model/vehicle_restrictions/vehicle_restriction_model.rs
@@ -4,7 +4,7 @@ use super::{
 };
 use routee_compass_core::model::{
     frontier::{frontier_model::FrontierModel, frontier_model_error::FrontierModelError},
-    property::edge::Edge,
+    road_network::Edge,
     state::state_model::StateModel,
     traversal::state::state_variable::StateVar,
 };
